Add fallbacks if the custom font doesn't load.
If the custom font doesn't load for some reason, the browser will fall back to some random font, possibly a serif font that looks Very Wrong. "system-ui" should choose the system font for the current platform, whatever it is, which is at least *a* system font even if it's not the right one for CS 1.6. For older browsers that don't support "system-ui", fall back to "sans-serif", which should look at least vaguely right.